# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/v1/car-model-years/": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Create a CarModelYear",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModelYears"
        ],
        "summary": "Create a CarModelYear",
        "parameters": [
          {
            "description": "Create a CarModelYear",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarModelYearRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "CarModelYear response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-model-years/get-by-filter": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get CarModelYears",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModelYears"
        ],
        "summary": "Get CarModelYears",
        "parameters": [
          {
            "description": "Request",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarModelYear response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarModelYearResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-model-years/{id}": {
      "get": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get a CarModelYear",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModelYears"
        ],
        "summary": "Get a CarModelYear",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "CarModelYear response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Update a CarModelYear",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModelYears"
        ],
        "summary": "Update a CarModelYear",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update a CarModelYear",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarModelYearRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarModelYear response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Delete a CarModelYear",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModelYears"
        ],
        "summary": "Delete a CarModelYear",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-models/": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Create a CarModel",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModels"
        ],
        "summary": "Create a CarModel",
        "parameters": [
          {
            "description": "Create a CarModel",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarModelRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "CarModel response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-models/get-by-filter": {
      "post": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get CarModels",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModels"
        ],
        "summary": "Get CarModels",
        "parameters": [
          {
            "description": "Request",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarModel response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarModelResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
    "/v1/car-models/{id}": {
      "get": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Get a CarModel",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModels"
        ],
        "summary": "Get a CarModel",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "CarModel response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Update a CarModel",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModels"
        ],
        "summary": "Update a CarModel",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update a CarModel",
            "name": "Request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarModelRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CarModel response",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "AuthBearer": []
          }
        ],
        "description": "Delete a CarModel",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "CarModels"
        ],
        "summary": "Delete a CarModel",
        "parameters": [
          {
            "type": "integer",
            "description": "Id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelColorResponse": {
      "type": "object",
      "properties": {
        "color": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.ColorResponse"
        },
        "id": {
          "type": "integer"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelCommentResponse": {
      "type": "object",
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.UserResponse"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelImageResponse": {
      "type": "object",
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "image": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.FileResponse"
        },
        "isMainImage": {
          "type": "boolean"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelPriceHistoryResponse": {
      "type": "object",
      "properties": {
        "carModelYearId": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "price": {
          "type": "number"
        },
        "priceAt": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelPropertyResponse": {
      "type": "object",
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "property": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PropertyResponse"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelResponse": {
      "type": "object",
      "properties": {
        "carModelColors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelColorResponse"
          }
        },
        "carModelComments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelCommentResponse"
          }
        },
        "carModelImages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelImageResponse"
          }
        },
        "carModelProperties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelPropertyResponse"
          }
        },
        "carModelYears": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse"
          }
        },
        "carType": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse"
        },
        "company": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse"
        },
        "gearbox": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.GearboxResponse"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse": {
      "type": "object",
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "carModelPriceHistories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelPriceHistoryResponse"
          }
        },
        "id": {
          "type": "integer"
        },
        "persianYear": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearWithoutDateResponse"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CarTypeResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.ColorResponse": {
      "type": "object",
      "properties": {
        "hexCode": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CompanyResponse": {
      "type": "object",
      "properties": {
        "country": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CountryResponse"
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarModelRequest": {
      "type": "object",
      "required": [
        "carTypeId",
        "companyId",
        "gearboxId",
        "name"
      ],
      "properties": {
        "carTypeId": {
          "type": "integer"
        },
        "companyId": {
          "type": "integer"
        },
        "gearboxId": {
          "type": "integer"
        },
        "name": {
          "type": "string",
          "maxLength": 15,
          "minLength": 3
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.CreateCarModelYearRequest": {
      "type": "object",
      "required": [
        "carModelId",
        "persianYearId"
      ],
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "persianYearId": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.FileResponse": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "directory": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "mimeType": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.GearboxResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.PersianYearWithoutDateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "persianTitle": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarModelRequest": {
      "type": "object",
      "properties": {
        "carTypeId": {
          "type": "integer"
        },
        "companyId": {
          "type": "integer"
        },
        "gearboxId": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_dto.UpdateCarModelYearRequest": {
      "type": "object",
      "properties": {
        "carModelId": {
          "type": "integer"
        },
        "persianYearId": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_api_dto.UserResponse": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "lastName": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_helper.BaseHttpResponse": {
      "type": "object",
      "properties": {
        "error": {},
        "result": {},
        "resultCode": {
          "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode"
        },
        "success": {
          "type": "boolean"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError"
          }
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_api_helper.ResultCode": {
      "type": "integer",
      "enum": [
        0,
        40001,
        40101,
        40301,
        40401,
        42901,
        42902,
        50001,
        50002
      ],
      "x-enum-varnames": [
        "Success",
        "ValidationError",
        "AuthError",
        "ForbiddenError",
        "NotFoundError",
        "LimiterError",
        "OtpLimiterError",
        "CustomRecovery",
        "InternalError"
      ]
    },
    "github_com_naeemaei_golang-clean-web-api_api_validation.ValidationError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "property": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_domain_filter.Filter": {
      "type": "object",
      "properties": {
        "filterType": {
          "description": "text number",
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "type": {
          "description": "contains notContains equals notEqual startsWith lessThan lessThanOrEqual greaterThan greaterThanOrEqual inRange endsWith",
          "type": "string"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarModelResponse": {
      "type": "object",
      "properties": {
        "hasNextPage": {
          "type": "boolean"
        },
        "hasPreviousPage": {
          "type": "boolean"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelResponse"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "totalPages": {
          "type": "integer"
        },
        "totalRows": {
          "type": "integer"
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PagedList-github_com_naeemaei_golang-clean-web-api_api_dto_CarModelYearResponse": {
      "type": "object",
      "properties": {
        "hasNextPage": {
          "type": "boolean"
        },
        "hasPreviousPage": {
          "type": "boolean"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_api_dto.CarModelYearResponse"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "totalPages": {
          "type": "integer"
        },
        "totalRows": {
          "type": "integer"
        }
      }
    },
...
    "github_com_naeemaei_golang-clean-web-api_domain_filter.PaginationInputWithFilter": {
      "type": "object",
      "properties": {
        "filter": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Filter"
          }
        },
        "pageNumber": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "sort": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_naeemaei_golang-clean-web-api_domain_filter.Sort"
          }
        }
      }
    },
    "github_com_naeemaei_golang-clean-web-api_domain_filter.Sort": {
      "type": "object",
      "properties": {
        "colId": {
          "type": "string"
        },
        "sort": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "AuthBearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}